server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.9
python-dotenv==1.0.0
httpx>=0.24.1,<0.25.0
Flask-Limiter==3.5.0
